+++ src/server/main/signals.py
@@ -1,32 +1,8 @@
-import string
-import secrets
-
-from django.db.models import signals, Max
+from django.db.models import signals
 from django.dispatch import receiver
-from django.contrib.auth.hashers import make_password
 
 from . import models
-
-
-@receiver([signals.pre_init], sender=models.User)
-def load_user_fixture(sender, *args, **kwargs):
-    if (
-        isinstance(kwargs.get("kwargs"), dict)
-        and kwargs["kwargs"].get("email")
-        and kwargs["kwargs"].get("id")  # Comment out when id is not serialized
-    ):
-        if not kwargs["kwargs"].get("id"):
-            kwargs["kwargs"]["id"] = (
-                models.User.objects.aggregate(Max("id")).get("id__max", 0) + 1
-            )
-
-        if not kwargs["kwargs"].get("password"):
-            password = "".join(
-                secrets.choice(string.ascii_letters + string.digits) for _ in range(10)
-            )
-            kwargs["kwargs"]["password"] = make_password(password)
-
-        print(f"User {kwargs['kwargs']['id']}: {kwargs['kwargs']['email']} {password}")
+from auth.models import User
 
 
 def user_created_hook(instance, *args, **kwargs):
@@ -51,7 +27,7 @@ def user_updated_hook(instance, *args, **kwargs):
     pass
 
 
-@receiver([signals.post_save], sender=models.User)
+@receiver([signals.post_save], sender=User)
 def user_post_save(sender, instance, *args, **kwargs):
     (
         user_created_hook
