+++ src/server/rest/serializers.py
@@ -1,10 +1,11 @@
 from datetime import timedelta
-from distutils.log import error
 from django.utils import timezone
 from django.core.exceptions import ValidationError
 from django.contrib.auth.password_validation import validate_password
 from rest_framework import serializers
+
 from main import models
+from auth.models import User
 
 
 class DynamicFieldsModelSerializer(serializers.ModelSerializer):
@@ -48,7 +49,7 @@ class UserSerializer(DynamicFieldsModelSerializer):
         ).data
 
     class Meta:
-        model = models.User
+        model = User
         fields = [
             "id",
             "email",
